requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
colorama==0.4.4
decorator==5.0.7
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
packaging==20.9
parso==0.8.2
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.17.0
selenium==3.141.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
